requirements.txt
absl-py==1.4.0
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.50
botocore==1.31.50
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.10.0
fastapi==0.103.1
filelock==3.12.4
fire==0.5.0
fonttools==4.42.1
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.2
idna==2.10
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
protobuf==4.24.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.6
rsa==4.9
s3transfer==0.6.2
sahi==0.11.14
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
supervision==0.14.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.181
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
yolov5==7.0.12
